Move app into its own package
