+++ internal/timecraft/run.go
@@ -11,6 +11,7 @@ import (
 const defaultFunction = "_start"
 
 func runModule(ctx context.Context, runtime wazero.Runtime, compiledModule wazero.CompiledModule, function string) error {
+	println("runModule")
 	if function == "" {
 		function = defaultFunction
 	}
@@ -27,6 +28,7 @@ func runModule(ctx context.Context, runtime wazero.Runtime, compiledModule wazer
 		return fmt.Errorf("function %q not found in guest", function)
 	}
 	_, err = fn.Call(ctx)
+	fmt.Println("fn.Call ended", err)
 	switch err {
 	case context.Canceled, context.DeadlineExceeded:
 		err = nil
