+++ sdk/python/src/timecraft/__init__.py
@@ -4,6 +4,8 @@ from .client import TaskState, TaskID
 from .client import HTTPRequest, HTTPResponse, Header
 from .client import ProcessID, ModuleSpec
 
+from .app import App, Entrypoint, Promise, Function
+
 from .server import serve_forever
 
 
@@ -12,4 +14,5 @@ __all__ = ['Client',
            'TaskState', 'TaskID',
            'HTTPRequest', 'HTTPResponse', 'Header',
            'ProcessID', 'ModuleSpec',
+           'App', 'Entrypoint', 'Promise', 'Function',
            'serve_forever']
