+++ testdata/python/app_test.py
@@ -1,33 +1,9 @@
-import asyncio
 import unittest
-import sys
-import pickle
+import asyncio
 
-from timecraft import App
+from .app import app
 
-
-app = App()
-
-class TestApp(unittest.TestCase):
+class TestApp(unittest.IsolatedAsyncioTestCase):
     def test_run(self):
-        asyncio.run(app.run(sys.argv[1:]))
-
-    @app.start()
-    async def main(self):
-        res = self.add.call(40, 2)
-    
-        while not await res.is_done():
-            await asyncio.sleep(0.1)
-    
-        resp = await res.result()
-        result = pickle.loads(resp.output.body)
-        self.assertEqual(result, 42)
-
-
-    @app.function(name="add")
-    def add(a, b):
-        return a + b
-
-
-if __name__ == "__main__":
-    unittest.main()
+        res = asyncio.run(app.run())
+        self.assertEqual(res, 42)
